Support depth-based image layer compositing.
Allows calling viz::ImageCudaDevice() with a depth format.

If fmt is a depth format, the image will be interpreted as a depth
image, and will be written to the depth buffer when rendering the
color image from a separate invocation of image_*() for the same
layer. This enables depth-compositing image layers with other Holoviz
layers. Supported depth formats are: D32_SFLOAT.

Implemented with an additional Vulkan pipeline for image + depth.